+++ goldens/public-api/router/index.md
@@ -650,7 +650,11 @@ export class RouteConfigLoadStart {
 
 // @public
 export class Router {
-    constructor(rootComponentType: Type<any> | null, urlSerializer: UrlSerializer, rootContexts: ChildrenOutletContexts, location: Location_2, injector: Injector, compiler: Compiler, config: Routes);
+    constructor(
+    rootComponentType: Type<any> | null,
+    urlSerializer: UrlSerializer,
+    rootContexts: ChildrenOutletContexts,
+    location: Location_2, injector: Injector, compiler: Compiler, config: Routes);
     // @deprecated
     canceledNavigationResolution: 'replace' | 'computed';
     // (undocumented)
