+++ goldens/size-tracking/integration-payloads.json
@@ -33,7 +33,7 @@
   "cli-hello-world-lazy": {
     "uncompressed": {
       "runtime": 2835,
-      "main": 225766,
+      "main": 226537,
       "polyfills": 33842,
       "src_app_lazy_lazy_routes_ts": 487
     }
