+++ packages/router/src/private_export.ts
@@ -9,7 +9,8 @@
 
 export {ɵEmptyOutletComponent} from './components/empty_outlet';
 export {DeprecatedLoadChildren as ɵDeprecatedLoadChildren} from './deprecated_load_children';
+export {RestoredState as ɵRestoredState} from './navigation_transition';
 export {withPreloading as ɵwithPreloading} from './provide_router';
-export {assignExtraOptionsToRouter as ɵassignExtraOptionsToRouter, RestoredState as ɵRestoredState} from './router';
+export {assignExtraOptionsToRouter as ɵassignExtraOptionsToRouter} from './router';
 export {ROUTER_PROVIDERS as ɵROUTER_PROVIDERS} from './router_module';
 export {flatten as ɵflatten} from './utils/collection';
