+++ packages/router/src/router.ts
@@ -8,162 +8,27 @@
 
 import {Location} from '@angular/common';
 import {Compiler, inject, Injectable, Injector, NgModuleRef, NgZone, Type, ɵConsole as Console, ɵRuntimeError as RuntimeError} from '@angular/core';
-import {BehaviorSubject, combineLatest, EMPTY, Observable, of, Subject, SubscriptionLike} from 'rxjs';
-import {catchError, defaultIfEmpty, filter, finalize, map, switchMap, take, tap} from 'rxjs/operators';
+import {BehaviorSubject, Observable, of, Subject, SubscriptionLike} from 'rxjs';
 
-import {createRouterState} from './create_router_state';
 import {createUrlTree} from './create_url_tree';
 import {RuntimeErrorCode} from './errors';
-import {Event, GuardsCheckEnd, GuardsCheckStart, NavigationCancel, NavigationCancellationCode, NavigationEnd, NavigationError, NavigationSkipped, NavigationSkippedCode, NavigationStart, NavigationTrigger, ResolveEnd, ResolveStart, RouteConfigLoadEnd, RouteConfigLoadStart, RoutesRecognized} from './events';
-import {NavigationBehaviorOptions, QueryParamsHandling, Route, Routes} from './models';
-import {isNavigationCancelingError, isRedirectingNavigationCancelingError, redirectingNavigationError} from './navigation_canceling_error';
-import {activateRoutes} from './operators/activate_routes';
-import {applyRedirects} from './operators/apply_redirects';
-import {checkGuards} from './operators/check_guards';
-import {recognize} from './operators/recognize';
-import {resolveData} from './operators/resolve_data';
-import {switchTap} from './operators/switch_tap';
+import {Event, NavigationCancel, NavigationCancellationCode, NavigationEnd, NavigationTrigger, RouteConfigLoadEnd, RouteConfigLoadStart} from './events';
+import {NavigationBehaviorOptions, Route, Routes} from './models';
+import {Navigation, NavigationExtras, NavigationTransition, NavigationTransitions, RestoredState, UrlCreationOptions} from './navigation_transition';
 import {TitleStrategy} from './page_title_strategy';
 import {RouteReuseStrategy} from './route_reuse_strategy';
 import {ErrorHandler, ExtraOptions, ROUTER_CONFIGURATION} from './router_config';
 import {RouterConfigLoader, ROUTES} from './router_config_loader';
 import {ChildrenOutletContexts} from './router_outlet_context';
-import {ActivatedRoute, ActivatedRouteSnapshot, createEmptyState, RouterState, RouterStateSnapshot} from './router_state';
+import {createEmptyState, RouterState} from './router_state';
 import {Params} from './shared';
 import {UrlHandlingStrategy} from './url_handling_strategy';
 import {containsTree, IsActiveMatchOptions, isUrlTree, UrlSerializer, UrlTree} from './url_tree';
 import {flatten} from './utils/collection';
 import {standardizeConfig, validateConfig} from './utils/config';
-import {Checks, getAllRouteGuards} from './utils/preactivation';
 
 
 const NG_DEV_MODE = typeof ngDevMode === 'undefined' || !!ngDevMode;
-
-/**
- * @description
- *
- * Options that modify the `Router` URL.
- * Supply an object containing any of these properties to a `Router` navigation function to
- * control how the target URL should be constructed.
- *
- * @see [Router.navigate() method](api/router/Router#navigate)
- * @see [Router.createUrlTree() method](api/router/Router#createurltree)
- * @see [Routing and Navigation guide](guide/router)
- *
- * @publicApi
- */
-export interface UrlCreationOptions {
-  /**
-   * Specifies a root URI to use for relative navigation.
-   *
-   * For example, consider the following route configuration where the parent route
-   * has two children.
-   *
-   * ```
-   * [{
-   *   path: 'parent',
-   *   component: ParentComponent,
-   *   children: [{
-   *     path: 'list',
-   *     component: ListComponent
-   *   },{
-   *     path: 'child',
-   *     component: ChildComponent
-   *   }]
-   * }]
-   * ```
-   *
-   * The following `go()` function navigates to the `list` route by
-   * interpreting the destination URI as relative to the activated `child`  route
-   *
-   * ```
-   *  @Component({...})
-   *  class ChildComponent {
-   *    constructor(private router: Router, private route: ActivatedRoute) {}
-   *
-   *    go() {
-   *      this.router.navigate(['../list'], { relativeTo: this.route });
-   *    }
-   *  }
-   * ```
-   *
-   * A value of `null` or `undefined` indicates that the navigation commands should be applied
-   * relative to the root.
-   */
-  relativeTo?: ActivatedRoute|null;
-
-  /**
-   * Sets query parameters to the URL.
-   *
-   * ```
-   * // Navigate to /results?page=1
-   * this.router.navigate(['/results'], { queryParams: { page: 1 } });
-   * ```
-   */
-  queryParams?: Params|null;
-
-  /**
-   * Sets the hash fragment for the URL.
-   *
-   * ```
-   * // Navigate to /results#top
-   * this.router.navigate(['/results'], { fragment: 'top' });
-   * ```
-   */
-  fragment?: string;
-
-  /**
-   * How to handle query parameters in the router link for the next navigation.
-   * One of:
-   * * `preserve` : Preserve current parameters.
-   * * `merge` : Merge new with current parameters.
-   *
-   * The "preserve" option discards any new query params:
-   * ```
-   * // from /view1?page=1 to/view2?page=1
-   * this.router.navigate(['/view2'], { queryParams: { page: 2 },  queryParamsHandling: "preserve"
-   * });
-   * ```
-   * The "merge" option appends new query params to the params from the current URL:
-   * ```
-   * // from /view1?page=1 to/view2?page=1&otherKey=2
-   * this.router.navigate(['/view2'], { queryParams: { otherKey: 2 },  queryParamsHandling: "merge"
-   * });
-   * ```
-   * In case of a key collision between current parameters and those in the `queryParams` object,
-   * the new value is used.
-   *
-   */
-  queryParamsHandling?: QueryParamsHandling|null;
-
-  /**
-   * When true, preserves the URL fragment for the next navigation
-   *
-   * ```
-   * // Preserve fragment from /results#top to /view#top
-   * this.router.navigate(['/view'], { preserveFragment: true });
-   * ```
-   */
-  preserveFragment?: boolean;
-}
-
-/**
- * @description
- *
- * Options that modify the `Router` navigation strategy.
- * Supply an object containing any of these properties to a `Router` navigation function to
- * control how the target URL should be constructed or interpreted.
- *
- * @see [Router.navigate() method](api/router/Router#navigate)
- * @see [Router.navigateByUrl() method](api/router/Router#navigatebyurl)
- * @see [Router.createUrlTree() method](api/router/Router#createurltree)
- * @see [Routing and Navigation guide](guide/router)
- * @see UrlCreationOptions
- * @see NavigationBehaviorOptions
- *
- * @publicApi
- */
-export interface NavigationExtras extends UrlCreationOptions, NavigationBehaviorOptions {}
 
 function defaultErrorHandler(error: any): any {
   throw error;
@@ -172,101 +37,6 @@ function defaultErrorHandler(error: any): any {
 function defaultMalformedUriErrorHandler(
     error: URIError, urlSerializer: UrlSerializer, url: string): UrlTree {
   return urlSerializer.parse('/');
-}
-
-export type RestoredState = {
-  [k: string]: any,
-  // TODO(#27607): Remove `navigationId` and `ɵrouterPageId` and move to `ng` or `ɵ` namespace.
-  navigationId: number,
-  // The `ɵ` prefix is there to reduce the chance of colliding with any existing user properties on
-  // the history state.
-  ɵrouterPageId?: number,
-};
-
-/**
- * Information about a navigation operation.
- * Retrieve the most recent navigation object with the
- * [Router.getCurrentNavigation() method](api/router/Router#getcurrentnavigation) .
- *
- * * *id* : The unique identifier of the current navigation.
- * * *initialUrl* : The target URL passed into the `Router#navigateByUrl()` call before navigation.
- * This is the value before the router has parsed or applied redirects to it.
- * * *extractedUrl* : The initial target URL after being parsed with `UrlSerializer.extract()`.
- * * *finalUrl* : The extracted URL after redirects have been applied.
- * This URL may not be available immediately, therefore this property can be `undefined`.
- * It is guaranteed to be set after the `RoutesRecognized` event fires.
- * * *trigger* : Identifies how this navigation was triggered.
- * -- 'imperative'--Triggered by `router.navigateByUrl` or `router.navigate`.
- * -- 'popstate'--Triggered by a popstate event.
- * -- 'hashchange'--Triggered by a hashchange event.
- * * *extras* : A `NavigationExtras` options object that controlled the strategy used for this
- * navigation.
- * * *previousNavigation* : The previously successful `Navigation` object. Only one previous
- * navigation is available, therefore this previous `Navigation` object has a `null` value for its
- * own `previousNavigation`.
- *
- * @publicApi
- */
-export interface Navigation {
-  /**
-   * The unique identifier of the current navigation.
-   */
-  id: number;
-  /**
-   * The target URL passed into the `Router#navigateByUrl()` call before navigation. This is
-   * the value before the router has parsed or applied redirects to it.
-   */
-  initialUrl: UrlTree;
-  /**
-   * The initial target URL after being parsed with `UrlHandlingStrategy.extract()`.
-   */
-  extractedUrl: UrlTree;
-  /**
-   * The extracted URL after redirects have been applied.
-   * This URL may not be available immediately, therefore this property can be `undefined`.
-   * It is guaranteed to be set after the `RoutesRecognized` event fires.
-   */
-  finalUrl?: UrlTree;
-  /**
-   * Identifies how this navigation was triggered.
-   *
-   * * 'imperative'--Triggered by `router.navigateByUrl` or `router.navigate`.
-   * * 'popstate'--Triggered by a popstate event.
-   * * 'hashchange'--Triggered by a hashchange event.
-   */
-  trigger: 'imperative'|'popstate'|'hashchange';
-  /**
-   * Options that controlled the strategy used for this navigation.
-   * See `NavigationExtras`.
-   */
-  extras: NavigationExtras;
-  /**
-   * The previously successful `Navigation` object. Only one previous navigation
-   * is available, therefore this previous `Navigation` object has a `null` value
-   * for its own `previousNavigation`.
-   */
-  previousNavigation: Navigation|null;
-}
-
-export interface NavigationTransition {
-  id: number;
-  targetPageId: number;
-  currentUrlTree: UrlTree;
-  extractedUrl: UrlTree;
-  urlAfterRedirects?: UrlTree;
-  rawUrl: UrlTree;
-  extras: NavigationExtras;
-  resolve: any;
-  reject: any;
-  promise: Promise<boolean>;
-  source: NavigationTrigger;
-  restoredState: RestoredState|null;
-  currentSnapshot: RouterStateSnapshot;
-  targetSnapshot: RouterStateSnapshot|null;
-  currentRouterState: RouterState;
-  targetRouterState: RouterState|null;
-  guards: Checks;
-  guardsResult: boolean|UrlTree|null;
 }
 
 /**
@@ -357,8 +127,9 @@ export class Router {
    *
    * This should match the `browserUrlTree` when a navigation succeeds. If the
    * `UrlHandlingStrategy.shouldProcessUrl` is `false`, only the `browserUrlTree` is updated.
+   * @internal
    */
-  private currentUrlTree: UrlTree;
+  currentUrlTree: UrlTree;
   /**
    * Meant to represent the entire browser url after a successful navigation. In the life of a
    * navigation transition:
@@ -369,6 +140,7 @@ export class Router {
    * rawUrlTree to be a combination of the urlAfterRedirects (again, this might only apply to part
    * of the initial url) and the rawUrl of the transition (which was the original navigation url in
    * its full form).
+   * @internal
    *
    * Note that this is _only_ here to support `UrlHandlingStrategy.extract` and
    * `UrlHandlingStrategy.shouldProcessUrl`. If those didn't exist, we could get by with
@@ -385,7 +157,7 @@ export class Router {
    * know what the browser's URL is for future navigations.
    *
    */
-  private rawUrlTree: UrlTree;
+  rawUrlTree: UrlTree;
   /**
    * Meant to represent the part of the browser url that the `Router` is set up to handle (via the
    * `UrlHandlingStrategy`). This value is updated immediately after the browser url is updated (or
@@ -400,16 +172,19 @@ export class Router {
    * activation with `'deferred'`.
    *
    * This should match the `currentUrlTree` when the navigation succeeds.
+   * @internal
    */
-  private browserUrlTree: UrlTree;
-  private readonly transitions: BehaviorSubject<NavigationTransition>;
+  browserUrlTree: UrlTree;
+  /** @internal */
+  readonly transitions: BehaviorSubject<NavigationTransition>;
   private navigations: Observable<NavigationTransition>;
-  private lastSuccessfulNavigation: Navigation|null = null;
-  private currentNavigation: Navigation|null = null;
+  /** @internal */
+  lastSuccessfulNavigation: Navigation|null = null;
   private disposed = false;
 
   private locationSubscription?: SubscriptionLike;
-  private navigationId: number = 0;
+  /** @internal */
+  navigationId: number = 0;
 
   /**
    * The id of the currently active page in the router.
@@ -428,8 +203,10 @@ export class Router {
   private get browserPageId(): number|undefined {
     return (this.location.getState() as RestoredState | null)?.ɵrouterPageId;
   }
-  private configLoader: RouterConfigLoader;
-  private ngModule: NgModuleRef<any>;
+  /** @internal */
+  configLoader: RouterConfigLoader;
+  /** @internal */
+  ngModule: NgModuleRef<any>;
   private console: Console;
   private isNgZoneEnabled: boolean = false;
 
@@ -581,14 +358,25 @@ export class Router {
    */
   canceledNavigationResolution: 'replace'|'computed' = 'replace';
 
+  private readonly navigationTransitions = new NavigationTransitions(this);
+
   /**
    * Creates the router service.
    */
   // TODO: vsavkin make internal after the final is out.
   constructor(
-      private rootComponentType: Type<any>|null, private urlSerializer: UrlSerializer,
-      private rootContexts: ChildrenOutletContexts, private location: Location, injector: Injector,
-      compiler: Compiler, public config: Routes) {
+      /** @internal */
+      public rootComponentType: Type<any>|null,
+      /** @internal */
+      readonly urlSerializer: UrlSerializer,
+      /** @internal */
+      readonly rootContexts: ChildrenOutletContexts,
+      /** @internal */
+      readonly location: Location,
+      injector: Injector,
+      compiler: Compiler,
+      public config: Routes,
+  ) {
     const onLoadStart = (r: Route) => this.triggerEvent(new RouteConfigLoadStart(r));
     const onLoadEnd = (r: Route) => this.triggerEvent(new RouteConfigLoadEnd(r));
     this.configLoader = injector.get(RouterConfigLoader);
@@ -627,402 +415,9 @@ export class Router {
       guards: {canActivateChecks: [], canDeactivateChecks: []},
       guardsResult: null,
     });
-    this.navigations = this.setupNavigations(this.transitions);
+    this.navigations = this.navigationTransitions.setupNavigations(this.transitions);
 
     this.processNavigations();
-  }
-
-  private setupNavigations(transitions: Observable<NavigationTransition>):
-      Observable<NavigationTransition> {
-    const eventsSubject = (this.events as Subject<Event>);
-    return transitions.pipe(
-               filter(t => t.id !== 0),
-
-               // Extract URL
-               map(t =>
-                       ({...t, extractedUrl: this.urlHandlingStrategy.extract(t.rawUrl)} as
-                        NavigationTransition)),
-
-               // Using switchMap so we cancel executing navigations when a new one comes in
-               switchMap(overallTransitionState => {
-                 let completed = false;
-                 let errored = false;
-                 return of(overallTransitionState)
-                     .pipe(
-                         // Store the Navigation object
-                         tap(t => {
-                           this.currentNavigation = {
-                             id: t.id,
-                             initialUrl: t.rawUrl,
-                             extractedUrl: t.extractedUrl,
-                             trigger: t.source,
-                             extras: t.extras,
-                             previousNavigation: this.lastSuccessfulNavigation ?
-                                 {...this.lastSuccessfulNavigation, previousNavigation: null} :
-                                 null
-                           };
-                         }),
-                         switchMap(t => {
-                           const browserUrlTree = this.browserUrlTree.toString();
-                           const urlTransition = !this.navigated ||
-                               t.extractedUrl.toString() !== browserUrlTree ||
-                               // Navigations which succeed or ones which fail and are cleaned up
-                               // correctly should result in `browserUrlTree` and `currentUrlTree`
-                               // matching. If this is not the case, assume something went wrong and
-                               // try processing the URL again.
-                               browserUrlTree !== this.currentUrlTree.toString();
-
-                           if (!urlTransition && this.onSameUrlNavigation !== 'reload') {
-                             const reason = NG_DEV_MODE ?
-                                 `Navigation to ${
-                                     t.rawUrl} was ignored because it is the same as the current Router URL.` :
-                                 '';
-                             this.triggerEvent(new NavigationSkipped(
-                                 t.id, this.serializeUrl(overallTransitionState.rawUrl), reason,
-                                 NavigationSkippedCode.IgnoredSameUrlNavigation));
-                             this.rawUrlTree = t.rawUrl;
-                             t.resolve(null);
-                             return EMPTY;
-                           }
-
-                           if (this.urlHandlingStrategy.shouldProcessUrl(t.rawUrl)) {
-                             // If the source of the navigation is from a browser event, the URL is
-                             // already updated. We already need to sync the internal state.
-                             if (isBrowserTriggeredNavigation(t.source)) {
-                               this.browserUrlTree = t.extractedUrl;
-                             }
-                             return of(t).pipe(
-                                 // Fire NavigationStart event
-                                 switchMap(t => {
-                                   const transition = this.transitions.getValue();
-                                   eventsSubject.next(new NavigationStart(
-                                       t.id, this.serializeUrl(t.extractedUrl), t.source,
-                                       t.restoredState));
-                                   if (transition !== this.transitions.getValue()) {
-                                     return EMPTY;
-                                   }
-
-                                   // This delay is required to match old behavior that forced
-                                   // navigation to always be async
-                                   return Promise.resolve(t);
-                                 }),
-
-                                 // ApplyRedirects
-                                 applyRedirects(
-                                     this.ngModule.injector, this.configLoader, this.urlSerializer,
-                                     this.config),
-
-                                 // Update the currentNavigation
-                                 // `urlAfterRedirects` is guaranteed to be set after this point
-                                 tap(t => {
-                                   this.currentNavigation = {
-                                     ...this.currentNavigation!,
-                                     finalUrl: t.urlAfterRedirects
-                                   };
-                                   overallTransitionState.urlAfterRedirects = t.urlAfterRedirects;
-                                 }),
-
-                                 // Recognize
-                                 recognize(
-                                     this.ngModule.injector, this.rootComponentType, this.config,
-                                     this.urlSerializer, this.paramsInheritanceStrategy),
-
-                                 // Update URL if in `eager` update mode
-                                 tap(t => {
-                                   overallTransitionState.targetSnapshot = t.targetSnapshot;
-                                   if (this.urlUpdateStrategy === 'eager') {
-                                     if (!t.extras.skipLocationChange) {
-                                       const rawUrl = this.urlHandlingStrategy.merge(
-                                           t.urlAfterRedirects!, t.rawUrl);
-                                       this.setBrowserUrl(rawUrl, t);
-                                     }
-                                     this.browserUrlTree = t.urlAfterRedirects!;
-                                   }
-
-                                   // Fire RoutesRecognized
-                                   const routesRecognized = new RoutesRecognized(
-                                       t.id, this.serializeUrl(t.extractedUrl),
-                                       this.serializeUrl(t.urlAfterRedirects!), t.targetSnapshot!);
-                                   eventsSubject.next(routesRecognized);
-                                 }));
-                           } else if (
-                               urlTransition &&
-                               this.urlHandlingStrategy.shouldProcessUrl(this.rawUrlTree)) {
-                             // When the current URL shouldn't be processed, but the previous one
-                             // was, we handle this by navigating from the current URL to an empty
-                             // state so deactivate guards can run.
-                             const {id, extractedUrl, source, restoredState, extras} = t;
-                             const navStart = new NavigationStart(
-                                 id, this.serializeUrl(extractedUrl), source, restoredState);
-                             eventsSubject.next(navStart);
-                             const targetSnapshot =
-                                 createEmptyState(extractedUrl, this.rootComponentType).snapshot;
-
-                             overallTransitionState = {
-                               ...t,
-                               targetSnapshot,
-                               urlAfterRedirects: extractedUrl,
-                               extras: {...extras, skipLocationChange: false, replaceUrl: false},
-                             };
-                             return of(overallTransitionState);
-                           } else {
-                             /* When neither the current or previous URL can be processed, do
-                              * nothing other than update router's internal reference to the
-                              * current "settled" URL. This way the next navigation will be coming
-                              * from the current URL in the browser.
-                              */
-                             const reason = NG_DEV_MODE ?
-                                 `Navigation was ignored because the UrlHandlingStrategy` +
-                                     ` indicated neither the current URL ${
-                                         this.rawUrlTree} nor target URL ${
-                                         t.rawUrl} should be processed.` :
-                                 '';
-                             this.triggerEvent(new NavigationSkipped(
-                                 t.id, this.serializeUrl(overallTransitionState.extractedUrl),
-                                 reason, NavigationSkippedCode.IgnoredByUrlHandlingStrategy));
-                             this.rawUrlTree = t.rawUrl;
-                             t.resolve(null);
-                             return EMPTY;
-                           }
-                         }),
-
-                         // --- GUARDS ---
-                         tap(t => {
-                           const guardsStart = new GuardsCheckStart(
-                               t.id, this.serializeUrl(t.extractedUrl),
-                               this.serializeUrl(t.urlAfterRedirects!), t.targetSnapshot!);
-                           this.triggerEvent(guardsStart);
-                         }),
-
-                         map(t => {
-                           overallTransitionState = {
-                             ...t,
-                             guards: getAllRouteGuards(
-                                 t.targetSnapshot!, t.currentSnapshot, this.rootContexts)
-                           };
-                           return overallTransitionState;
-                         }),
-
-                         checkGuards(
-                             this.ngModule.injector, (evt: Event) => this.triggerEvent(evt)),
-                         tap(t => {
-                           overallTransitionState.guardsResult = t.guardsResult;
-                           if (isUrlTree(t.guardsResult)) {
-                             throw redirectingNavigationError(this.urlSerializer, t.guardsResult);
-                           }
-
-                           const guardsEnd = new GuardsCheckEnd(
-                               t.id, this.serializeUrl(t.extractedUrl),
-                               this.serializeUrl(t.urlAfterRedirects!), t.targetSnapshot!,
-                               !!t.guardsResult);
-                           this.triggerEvent(guardsEnd);
-                         }),
-
-                         filter(t => {
-                           if (!t.guardsResult) {
-                             this.restoreHistory(t);
-                             this.cancelNavigationTransition(
-                                 t, '', NavigationCancellationCode.GuardRejected);
-                             return false;
-                           }
-                           return true;
-                         }),
-
-                         // --- RESOLVE ---
-                         switchTap(t => {
-                           if (t.guards.canActivateChecks.length) {
-                             return of(t).pipe(
-                                 tap(t => {
-                                   const resolveStart = new ResolveStart(
-                                       t.id, this.serializeUrl(t.extractedUrl),
-                                       this.serializeUrl(t.urlAfterRedirects!), t.targetSnapshot!);
-                                   this.triggerEvent(resolveStart);
-                                 }),
-                                 switchMap(t => {
-                                   let dataResolved = false;
-                                   return of(t).pipe(
-                                       resolveData(
-                                           this.paramsInheritanceStrategy, this.ngModule.injector),
-                                       tap({
-                                         next: () => dataResolved = true,
-                                         complete: () => {
-                                           if (!dataResolved) {
-                                             this.restoreHistory(t);
-                                             this.cancelNavigationTransition(
-                                                 t,
-                                                 NG_DEV_MODE ?
-                                                     `At least one route resolver didn't emit any value.` :
-                                                     '',
-                                                 NavigationCancellationCode.NoDataFromResolver);
-                                           }
-                                         }
-                                       }),
-                                   );
-                                 }),
-                                 tap(t => {
-                                   const resolveEnd = new ResolveEnd(
-                                       t.id, this.serializeUrl(t.extractedUrl),
-                                       this.serializeUrl(t.urlAfterRedirects!), t.targetSnapshot!);
-                                   this.triggerEvent(resolveEnd);
-                                 }));
-                           }
-                           return undefined;
-                         }),
-
-                         // --- LOAD COMPONENTS ---
-                         switchTap((t: NavigationTransition) => {
-                           const loadComponents =
-                               (route: ActivatedRouteSnapshot): Array<Observable<void>> => {
-                                 const loaders: Array<Observable<void>> = [];
-                                 if (route.routeConfig?.loadComponent &&
-                                     !route.routeConfig._loadedComponent) {
-                                   loaders.push(this.configLoader.loadComponent(route.routeConfig)
-                                                    .pipe(
-                                                        tap(loadedComponent => {
-                                                          route.component = loadedComponent;
-                                                        }),
-                                                        map(() => void 0),
-                                                        ));
-                                 }
-                                 for (const child of route.children) {
-                                   loaders.push(...loadComponents(child));
-                                 }
-                                 return loaders;
-                               };
-                           return combineLatest(loadComponents(t.targetSnapshot!.root))
-                               .pipe(defaultIfEmpty(), take(1));
-                         }),
-
-                         switchTap(() => this.afterPreactivation()),
-
-                         map((t: NavigationTransition) => {
-                           const targetRouterState = createRouterState(
-                               this.routeReuseStrategy, t.targetSnapshot!, t.currentRouterState);
-                           overallTransitionState = {...t, targetRouterState};
-                           return (overallTransitionState);
-                         }),
-
-                         /* Once here, we are about to activate synchronously. The assumption is
-                            this will succeed, and user code may read from the Router service.
-                            Therefore before activation, we need to update router properties storing
-                            the current URL and the RouterState, as well as updated the browser URL.
-                            All this should happen *before* activating. */
-                         tap((t: NavigationTransition) => {
-                           this.currentUrlTree = t.urlAfterRedirects!;
-                           this.rawUrlTree =
-                               this.urlHandlingStrategy.merge(t.urlAfterRedirects!, t.rawUrl);
-
-                           (this as {routerState: RouterState}).routerState = t.targetRouterState!;
-
-                           if (this.urlUpdateStrategy === 'deferred') {
-                             if (!t.extras.skipLocationChange) {
-                               this.setBrowserUrl(this.rawUrlTree, t);
-                             }
-                             this.browserUrlTree = t.urlAfterRedirects!;
-                           }
-                         }),
-
-                         activateRoutes(
-                             this.rootContexts, this.routeReuseStrategy,
-                             (evt: Event) => this.triggerEvent(evt)),
-
-                         tap({
-                           next() {
-                             completed = true;
-                           },
-                           complete() {
-                             completed = true;
-                           }
-                         }),
-                         finalize(() => {
-                           /* When the navigation stream finishes either through error or success,
-                            * we set the `completed` or `errored` flag. However, there are some
-                            * situations where we could get here without either of those being set.
-                            * For instance, a redirect during NavigationStart. Therefore, this is a
-                            * catch-all to make sure the NavigationCancel event is fired when a
-                            * navigation gets cancelled but not caught by other means. */
-                           if (!completed && !errored) {
-                             const cancelationReason = NG_DEV_MODE ?
-                                 `Navigation ID ${
-                                     overallTransitionState
-                                         .id} is not equal to the current navigation id ${
-                                     this.navigationId}` :
-                                 '';
-                             this.cancelNavigationTransition(
-                                 overallTransitionState, cancelationReason,
-                                 NavigationCancellationCode.SupersededByNewNavigation);
-                           }
-                           // Only clear current navigation if it is still set to the one that
-                           // finalized.
-                           if (this.currentNavigation?.id === overallTransitionState.id) {
-                             this.currentNavigation = null;
-                           }
-                         }),
-                         catchError((e) => {
-                           errored = true;
-                           /* This error type is issued during Redirect, and is handled as a
-                            * cancellation rather than an error. */
-                           if (isNavigationCancelingError(e)) {
-                             if (!isRedirectingNavigationCancelingError(e)) {
-                               // Set property only if we're not redirecting. If we landed on a page
-                               // and redirect to `/` route, the new navigation is going to see the
-                               // `/` isn't a change from the default currentUrlTree and won't
-                               // navigate. This is only applicable with initial navigation, so
-                               // setting `navigated` only when not redirecting resolves this
-                               // scenario.
-                               this.navigated = true;
-                               this.restoreHistory(overallTransitionState, true);
-                             }
-                             const navCancel = new NavigationCancel(
-                                 overallTransitionState.id,
-                                 this.serializeUrl(overallTransitionState.extractedUrl), e.message,
-                                 e.cancellationCode);
-                             eventsSubject.next(navCancel);
-
-                             // When redirecting, we need to delay resolving the navigation
-                             // promise and push it to the redirect navigation
-                             if (!isRedirectingNavigationCancelingError(e)) {
-                               overallTransitionState.resolve(false);
-                             } else {
-                               const mergedTree =
-                                   this.urlHandlingStrategy.merge(e.url, this.rawUrlTree);
-                               const extras = {
-                                 skipLocationChange:
-                                     overallTransitionState.extras.skipLocationChange,
-                                 // The URL is already updated at this point if we have 'eager' URL
-                                 // updates or if the navigation was triggered by the browser (back
-                                 // button, URL bar, etc). We want to replace that item in history
-                                 // if the navigation is rejected.
-                                 replaceUrl: this.urlUpdateStrategy === 'eager' ||
-                                     isBrowserTriggeredNavigation(overallTransitionState.source)
-                               };
-
-                               this.scheduleNavigation(mergedTree, 'imperative', null, extras, {
-                                 resolve: overallTransitionState.resolve,
-                                 reject: overallTransitionState.reject,
-                                 promise: overallTransitionState.promise
-                               });
-                             }
-
-                             /* All other errors should reset to the router's internal URL reference
-                              * to the pre-error state. */
-                           } else {
-                             this.restoreHistory(overallTransitionState, true);
-                             const navError = new NavigationError(
-                                 overallTransitionState.id,
-                                 this.serializeUrl(overallTransitionState.extractedUrl), e,
-                                 overallTransitionState.targetSnapshot ?? undefined);
-                             eventsSubject.next(navError);
-                             try {
-                               overallTransitionState.resolve(this.errorHandler(e));
-                             } catch (ee) {
-                               overallTransitionState.reject(ee);
-                             }
-                           }
-                           return EMPTY;
-                         }));
-                 // TODO(jasonaden): remove cast once g3 is on updated TypeScript
-               })) as any as Observable<NavigationTransition>;
   }
 
   /**
@@ -1107,7 +502,7 @@ export class Router {
    * and `null` when idle.
    */
   getCurrentNavigation(): Navigation|null {
-    return this.currentNavigation;
+    return this.navigationTransitions.currentNavigation;
   }
 
   /** @internal */
@@ -1368,7 +763,7 @@ export class Router {
           (this.events as Subject<Event>)
               .next(new NavigationEnd(
                   t.id, this.serializeUrl(t.extractedUrl), this.serializeUrl(this.currentUrlTree)));
-          this.lastSuccessfulNavigation = this.currentNavigation;
+          this.lastSuccessfulNavigation = this.getCurrentNavigation();
           this.titleStrategy?.updateTitle(this.routerState.snapshot);
           t.resolve(true);
         },
@@ -1377,7 +772,8 @@ export class Router {
         });
   }
 
-  private scheduleNavigation(
+  /** @internal */
+  scheduleNavigation(
       rawUrl: UrlTree, source: NavigationTrigger, restoredState: RestoredState|null,
       extras: NavigationExtras,
       priorPromise?: {resolve: any, reject: any, promise: Promise<boolean>}): Promise<boolean> {
@@ -1447,10 +843,14 @@ export class Router {
     });
   }
 
-  private setBrowserUrl(url: UrlTree, t: NavigationTransition) {
+  /** @internal */
+  setBrowserUrl(url: UrlTree, transition: NavigationTransition) {
     const path = this.urlSerializer.serialize(url);
-    const state = {...t.extras.state, ...this.generateNgRouterState(t.id, t.targetPageId)};
-    if (this.location.isCurrentPathEqualTo(path) || !!t.extras.replaceUrl) {
+    const state = {
+      ...transition.extras.state,
+      ...this.generateNgRouterState(transition.id, transition.targetPageId)
+    };
+    if (this.location.isCurrentPathEqualTo(path) || !!transition.extras.replaceUrl) {
       this.location.replaceState(path, '', state);
     } else {
       this.location.go(path, '', state);
@@ -1460,29 +860,31 @@ export class Router {
   /**
    * Performs the necessary rollback action to restore the browser URL to the
    * state before the transition.
+   * @internal
    */
-  private restoreHistory(t: NavigationTransition, restoringFromCaughtError = false) {
+  restoreHistory(transition: NavigationTransition, restoringFromCaughtError = false) {
     if (this.canceledNavigationResolution === 'computed') {
-      const targetPagePosition = this.currentPageId - t.targetPageId;
+      const targetPagePosition = this.currentPageId - transition.targetPageId;
       // The navigator change the location before triggered the browser event,
       // so we need to go back to the current url if the navigation is canceled.
       // Also, when navigation gets cancelled while using url update strategy eager, then we need to
       // go back. Because, when `urlUpdateStrategy` is `eager`; `setBrowserUrl` method is called
       // before any verification.
       const browserUrlUpdateOccurred =
-          (t.source === 'popstate' || this.urlUpdateStrategy === 'eager' ||
-           this.currentUrlTree === this.currentNavigation?.finalUrl);
+          (transition.source === 'popstate' || this.urlUpdateStrategy === 'eager' ||
+           this.currentUrlTree === this.getCurrentNavigation()?.finalUrl);
       if (browserUrlUpdateOccurred && targetPagePosition !== 0) {
         this.location.historyGo(targetPagePosition);
       } else if (
-          this.currentUrlTree === this.currentNavigation?.finalUrl && targetPagePosition === 0) {
+          this.currentUrlTree === this.getCurrentNavigation()?.finalUrl &&
+          targetPagePosition === 0) {
         // We got to the activation stage (where currentUrlTree is set to the navigation's
         // finalUrl), but we weren't moving anywhere in history (skipLocationChange or replaceUrl).
         // We still need to reset the router state back to what it was when the navigation started.
-        this.resetState(t);
+        this.resetState(transition);
         // TODO(atscott): resetting the `browserUrlTree` should really be done in `resetState`.
         // Investigate if this can be done by running TGP.
-        this.browserUrlTree = t.currentUrlTree;
+        this.browserUrlTree = transition.currentUrlTree;
         this.resetUrlToCurrentUrlTree();
       } else {
         // The browser URL and router state was not updated before the navigation cancelled so
@@ -1494,7 +896,7 @@ export class Router {
       // reject. For 'eager' navigations, it seems like we also really should reset the state
       // because the navigation was cancelled. Investigate if this can be done by running TGP.
       if (restoringFromCaughtError) {
-        this.resetState(t);
+        this.resetState(transition);
       }
       this.resetUrlToCurrentUrlTree();
     }
@@ -1517,11 +919,13 @@ export class Router {
         this.generateNgRouterState(this.lastSuccessfulId, this.currentPageId));
   }
 
-  private cancelNavigationTransition(
-      t: NavigationTransition, reason: string, code: NavigationCancellationCode) {
-    const navCancel = new NavigationCancel(t.id, this.serializeUrl(t.extractedUrl), reason, code);
+  /** @internal */
+  cancelNavigationTransition(
+      transition: NavigationTransition, reason: string, code: NavigationCancellationCode) {
+    const navCancel = new NavigationCancel(
+        transition.id, this.serializeUrl(transition.extractedUrl), reason, code);
     this.triggerEvent(navCancel);
-    t.resolve(false);
+    transition.resolve(false);
   }
 
   private generateNgRouterState(navigationId: number, routerPageId?: number) {
@@ -1541,8 +945,4 @@ function validateCommands(commands: string[]): void {
           NG_DEV_MODE && `The requested path contains ${cmd} segment at index ${i}`);
     }
   }
-}
-
-function isBrowserTriggeredNavigation(source: 'imperative'|'popstate'|'hashchange') {
-  return source !== 'imperative';
 }
